requirements.txt
argon2-cffi==19.1.0
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
channels==2.1.3
constantly==15.1.0
daphne==2.2.2
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
mod-wsgi==4.6.5
mysqlclient==1.4.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.1
pycparser==2.18
PyHamcrest==1.9.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
Twisted==18.7.0
txaio==18.8.1
whitenoise==4.1.2
zope.interface==4.5.0
